drivers: hse: add RNG support
Implement the HSE RNG driver and override the API calls for getting
random data. Disable the software PRNG when the HSE driver is in use.

The HSE RNG driver will store random data in a cache-like structure and
refill it using asynchronous requests. When the upper layers request
random data, the driver retrieves it from cache. If the random data
falls under a set threshold, an asynchronous request is sent to HSE
to refill the cache.

Issue: ALB-8947, ALB-9304
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>