drivers: hse: add buffer management utilities
Buffers used for data transfers between the host and HSE Firmware must
be aligned to the cacheline's size and occupy whole cache line(s)
because certain cache operations will be performed on these buffers.
Having them isolated on a single or multiple cache lines of their own
helps to keep the data transfers to/from HSE in sync with the
main memory.

Besides memory alignment, management of the buffers' physical addresses
must be done as HSE Firmware operates on physical ranges.

This commit introduces the hse_buf struct containg useful information
for dealing with a buffer's requirements.

Issue: ALB-8988, ALB-9301
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>