drivers: hse: add shared secret and hmac key rings
The shared secret key slots and hmac key slots are necessary for key
derivation operations. Add CFG_* parameters for each of them and
initialize their key rings in hse_core.

Also, reduce the number of AES Key Slots: there is a total of 20 keys
that can be formatted in the catalog. In order to add a shared secret
key slot, one AES key slot had to be removed.

Issue: ALB-8819
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>