plat-s32: re-initialize the console on the resume path
In the scenario of suspend & resume, ATF does not configures the console
on the resume path, unless BL31 logs are enabled (BL31 logs are disabled
by default). The console is in an undefined state after the platform
wakes up, making it unusable wihtout re-configuration. This leads to the
platform's secondary cores hanging while trying to use the console in
optee.

Go through the initialization routine on the resume path.

Issue: ALB-9306
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>