drivers: hse: add HSE_FWDIR variable to include HSE interface headers
The newly introduced build parameter HSE_FWDIR contains the path to the
HSE Firmware Package, as provided by NXP. The path must be set by the
user at build time, otherwise it would resort to a default one.

The HSE interface headers are now included in the build. Up until now
OP-TEE would replicate the interface's structs, types and definitions
mainly because the Linux HSE driver, which was the basis for the OP-TEE
driver, would do so. However, as opposed to Linux, OP-TEE's license
does not contaminate HSE Interface's. Hence, the HSE Interface is now
included and is meant to be used throughout the HSE Driver to ease
development.

Issue: ALB-9415
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>