drivers: hse: add aes-ctr mode in cipher driver
AES Counter mode is not a NOPAD block mode (as opposed to ECB or CBC),
meaning that the driver might receive the plaintext in chunks
of data (through the update routine) which are not necessarily the size
of an AES block size. In this regard, the context must keep track of
the last processed block from a previous update so that it would be
prepended to the current input and generate the expected ciphertext.

Issue: ALB-9903
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>