drivers: hse: refactor key management
Renamed the 'key_ring' structures/functions to 'key_group' to better
reflect the abstraction of HSE's key groups.

The key groups are now kept in a list data structure instead of
individual arrays per key type. Each entry in the list represents a
key group and there's a unique key group per key type in the list. This
change makes the addition of new key groups far easier and more scalable.

The key slot acquisition now returns a key handle. The key handle is the
unique identifier of a key slot in HSE's internal RAM/NVM Catalogs.
The previous implementation returned a 'struct hse_key' which also
contained the key handle of the slot and other fields used by the key
management functions. However, the consumers of the key slots only
need to use the key handles when filling service descriptors.

Issue: ALB-10053
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>