drivers: hse: HSE Hash driver support
Introduce support for digest algorithms
SHA-{1, 224, 256, 384, 512} and MD5 in
the HSE hash driver. To avoid performance
loss, instead of sending one block at a
time(sending all of them at the same is
impossible, the inputs are too big for
OP-TEE to handle), a safety net was
introduced, sending 20 blocks at a time,
and the remaining data in the final operation.
This way, for smaller blocks the throughtput
is ~2 times better, and for bigger blocks ~16
times better. Not choosing this buffered approach
would result in a driver slower than the previous
software approach.

Issue: ALB-8816
Signed-off-by: David Anghel <david.anghel@nxp.com>