drivers: hse: remove unnecessary decryption buffer resizing
The previous resizing of the cipher buffer (the input buffer in case
of decryption) is not necessary anymore. Its initial purpose was to
avoid input buffers that would have had a size greater than 512 bytes
because the standard firmware only supports a maximum size of 256 bytes
for RSA Keys.

Issue: ALB-10454
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>