drivers: hse: Add support for ECDH in the HSE driver
To use ECDH, you must enable it using a special flag
(default value is disabled) `CFG_NXP_HSE_ECDH_DRV=y`.
Because HSE computes ECDH internally and saves it inside
a SHARED_SECRET slot, without the ability to export it,
the approach is to do the DeriveCopy operation and then
export it. But, since HSE does not allow keys to be exported
in plain, the output is encrypted with AES-256-CTR with a
zero-ed IV and key. The normal tests inside xtest will not
produce the desired output, other test vectors that have
encrypted output were introduced. If the flag is disabled,
ECDH will not work at all, it will just return TEE_SUCCESS.

Issue: ALB-9957
Signed-off-by: David Anghel <david.anghel@nxp.com>