drivers: hse: Enable CFG_CORE_BGET_BESTFIT
Enabling this flag improves memory alignment,
with the cost of a higher allocation time.
The temporal overhead is not visible when
using anything inside the HSE driver, so
enabling this flag will not be costly.
By enabling this, we can also increase
the safety-net of the hash operation to
increase performance.

Issue: ALB-10797
Signed-off-by: David Anghel <david.anghel@nxp.com>